[clang] Split ObjectFilePCHContainerReader from ObjectFilePCHContainerWriter
Close https://github.com/llvm/llvm-project/issues/99479

See https://github.com/llvm/llvm-project/issues/99479 for details
