+++ clang/lib/Serialization/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(LLVM_LINK_COMPONENTS
   BitReader
   BitstreamReader
+  Object
   Support
   TargetParser
   )
@@ -21,6 +22,7 @@ add_clang_library(clangSerialization
   ModuleFileExtension.cpp
   ModuleManager.cpp
   PCHContainerOperations.cpp
+  ObjectFilePCHContainerReader.cpp
 
   ADDITIONAL_HEADERS
   ASTCommon.h
