+++ clang/tools/c-index-test/CMakeLists.txt
@@ -27,7 +27,6 @@ else()
     libclang
     clangAST
     clangBasic
-    clangCodeGen
     clangFrontend
     clangIndex
     clangSerialization
